consider error cases for joining room
