+++ src/main/java/com/wafflestudio/draft/dto/request/GetRegionsRequest.java
@@ -1,0 +1,8 @@
+package com.wafflestudio.draft.dto.request;
+
+import lombok.Data;
+
+@Data
+public class GetRegionsRequest {
+    private String name;
+}
